+++ src/main/java/com/kard/api/resources/notifications/types/EarnedRewardSettledAttributes.java
@@ -13,6 +13,7 @@ import com.fasterxml.jackson.annotation.Nulls;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.kard.api.core.ObjectMappers;
 import com.kard.api.resources.commons.types.CommissionValue;
+import java.time.OffsetDateTime;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
@@ -34,6 +35,8 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
 
     private final CommissionValue commissionEarned;
 
+    private final Optional<OffsetDateTime> transactionTimestamp;
+
     private final Map<String, Object> additionalProperties;
 
     private EarnedRewardSettledAttributes(
@@ -43,6 +46,7 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
             Optional<String> surveyUrl,
             Optional<String> cardProductId,
             CommissionValue commissionEarned,
+            Optional<OffsetDateTime> transactionTimestamp,
             Map<String, Object> additionalProperties) {
         this.message = message;
         this.name = name;
@@ -50,6 +54,7 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
         this.surveyUrl = surveyUrl;
         this.cardProductId = cardProductId;
         this.commissionEarned = commissionEarned;
+        this.transactionTimestamp = transactionTimestamp;
         this.additionalProperties = additionalProperties;
     }
 
@@ -103,6 +108,14 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
         return commissionEarned;
     }
 
+    /**
+     * @return The timestamp of the originating transaction in ISO format
+     */
+    @JsonProperty("transactionTimestamp")
+    public Optional<OffsetDateTime> getTransactionTimestamp() {
+        return transactionTimestamp;
+    }
+
     @java.lang.Override
     public boolean equals(Object other) {
         if (this == other) return true;
@@ -120,7 +133,8 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
                 && attributionUrl.equals(other.attributionUrl)
                 && surveyUrl.equals(other.surveyUrl)
                 && cardProductId.equals(other.cardProductId)
-                && commissionEarned.equals(other.commissionEarned);
+                && commissionEarned.equals(other.commissionEarned)
+                && transactionTimestamp.equals(other.transactionTimestamp);
     }
 
     @java.lang.Override
@@ -131,7 +145,8 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
                 this.attributionUrl,
                 this.surveyUrl,
                 this.cardProductId,
-                this.commissionEarned);
+                this.commissionEarned,
+                this.transactionTimestamp);
     }
 
     @java.lang.Override
@@ -190,6 +205,13 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
         _FinalStage cardProductId(Optional<String> cardProductId);
 
         _FinalStage cardProductId(String cardProductId);
+
+        /**
+         * <p>The timestamp of the originating transaction in ISO format</p>
+         */
+        _FinalStage transactionTimestamp(Optional<OffsetDateTime> transactionTimestamp);
+
+        _FinalStage transactionTimestamp(OffsetDateTime transactionTimestamp);
     }
 
     @JsonIgnoreProperties(ignoreUnknown = true)
@@ -202,6 +224,8 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
         private String attributionUrl;
 
         private CommissionValue commissionEarned;
+
+        private Optional<OffsetDateTime> transactionTimestamp = Optional.empty();
 
         private Optional<String> cardProductId = Optional.empty();
 
@@ -220,6 +244,7 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
             surveyUrl(other.getSurveyUrl());
             cardProductId(other.getCardProductId());
             commissionEarned(other.getCommissionEarned());
+            transactionTimestamp(other.getTransactionTimestamp());
             return this;
         }
 
@@ -267,6 +292,26 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
         }
 
         /**
+         * <p>The timestamp of the originating transaction in ISO format</p>
+         * @return Reference to {@code this} so that method calls can be chained together.
+         */
+        @java.lang.Override
+        public _FinalStage transactionTimestamp(OffsetDateTime transactionTimestamp) {
+            this.transactionTimestamp = Optional.ofNullable(transactionTimestamp);
+            return this;
+        }
+
+        /**
+         * <p>The timestamp of the originating transaction in ISO format</p>
+         */
+        @java.lang.Override
+        @JsonSetter(value = "transactionTimestamp", nulls = Nulls.SKIP)
+        public _FinalStage transactionTimestamp(Optional<OffsetDateTime> transactionTimestamp) {
+            this.transactionTimestamp = transactionTimestamp;
+            return this;
+        }
+
+        /**
          * <p>The ID of the card product</p>
          * @return Reference to {@code this} so that method calls can be chained together.
          */
@@ -309,7 +354,14 @@ public final class EarnedRewardSettledAttributes implements IRewardNotificationA
         @java.lang.Override
         public EarnedRewardSettledAttributes build() {
             return new EarnedRewardSettledAttributes(
-                    message, name, attributionUrl, surveyUrl, cardProductId, commissionEarned, additionalProperties);
+                    message,
+                    name,
+                    attributionUrl,
+                    surveyUrl,
+                    cardProductId,
+                    commissionEarned,
+                    transactionTimestamp,
+                    additionalProperties);
         }
 
         @java.lang.Override